refactor(udgeom): remove 443 lines of dead code (G26), share mesh utils (G27), fix ptp (G25)
- G26: delete 15 confirmed-dead functions from geometry_generation.py
  (1797->1354 lines, ~25%): the _triangulate_simple_polygon / _refine_skewed_
  triangles / _polygon_to_triangle_pslg chains, _triangles_to_mesh,
  _sample_linestring_points, _rings_to_triangle_pslg, _ordered_rectangle_
  boundary_points, _sample_interval; plus an unused reduce import, a discarded
  footprint_polys block in create_canyons, and a dead tol==0 block in create_cubes.
  Each grep-confirmed referenced nowhere but its own (removed) cluster.
- G27: extract the byte-identical _iter_polygons / _project_vertical_face /
  _copy_mesh into new udgeom/_meshutil.py; check_mesh/fix_mesh/truncate_below_ground
  import from it. (geometry_generation's same-named nested helper had diverged and
  was left untouched.)
- G25: fix ndarray.ptp() -> np.ptp() (removed in NumPy 2.0) in the SciPy-less
  ground-triangulation fallback.
Protected by the udgeom golden/MATLAB-reference suite: 86 udgeom + 10 meshgraph
tests unchanged and green.
