+++ tools/python/udgeom/fix_mesh.py
@@ -31,50 +31,17 @@ from .check_mesh import (
 )
 from .delete_ground import delete_ground
 from ._meshgraph import build_face_adjacency, connected_components
+from ._meshutil import _copy_mesh, _iter_polygons, _project_vertical_face
 
 from shapely.geometry import GeometryCollection, MultiPolygon, Polygon
 from shapely.ops import triangulate, unary_union
 import triangle as triangle_lib
 
 
-def _copy_mesh(mesh: "trimesh.Trimesh") -> "trimesh.Trimesh":
-    return trimesh.Trimesh(
-        vertices=np.asarray(mesh.vertices, dtype=float).copy(),
-        faces=np.asarray(mesh.faces, dtype=int).copy(),
-        process=False,
-    )
-
-
 def _face_components(mesh: "trimesh.Trimesh") -> list[np.ndarray]:
     if len(mesh.faces) == 0:
         return []
     return connected_components(build_face_adjacency(mesh))
-
-
-def _iter_polygons(geom):
-    if geom.is_empty:
-        return
-    if isinstance(geom, Polygon):
-        yield geom
-    elif isinstance(geom, MultiPolygon):
-        for poly in geom.geoms:
-            yield from _iter_polygons(poly)
-    elif isinstance(geom, GeometryCollection):
-        for item in geom.geoms:
-            yield from _iter_polygons(item)
-
-
-def _project_vertical_face(vertices: np.ndarray, axis: int) -> Polygon | None:
-    if axis == 0:
-        coords = vertices[:, [1, 2]]
-    else:
-        coords = vertices[:, [0, 2]]
-    poly = Polygon(coords)
-    if not poly.is_valid:
-        poly = poly.buffer(0)
-    if poly.is_empty or poly.area <= 1.0e-12:
-        return None
-    return poly
 
 
 def _triangle_to_3d(coords2d: np.ndarray, axis: int, plane_value: float, sign: int) -> np.ndarray:
